revise: move API entrypoints to wdl crate
