+++ wdl-analysis/Cargo.toml
@@ -12,7 +12,6 @@ documentation = "https://docs.rs/wdl-analysis"
 
 [dependencies]
 wdl-ast = { path = "../wdl-ast", version = "0.9.0" }
-wdl-lint = { path = "../wdl-lint", version = "0.8.0" }
 wdl-format = { path = "../wdl-format", version = "0.3.0" }
 
 anyhow = { workspace = true }
@@ -26,7 +25,6 @@ petgraph = { workspace = true }
 futures = { workspace = true }
 path-clean = { workspace = true }
 indexmap = { workspace = true }
-indicatif = { workspace = true }
 line-index = { workspace = true }
 uuid = { workspace = true, features = ["v4"] }
 walkdir = { workspace = true }
