+++ wdl/src/bin/wdl.rs
@@ -24,16 +24,16 @@ use codespan_reporting::term::termcolor::ColorChoice;
 use codespan_reporting::term::termcolor::StandardStream;
 use colored::Colorize;
 use tracing_log::AsTrace;
+use wdl::analyze;
 use wdl::ast::Diagnostic;
 use wdl::ast::Document;
 use wdl::ast::Validator;
 use wdl::lint::LintVisitor;
-use wdl_analysis::analyze;
+use wdl::run;
+use wdl::validate_inputs;
 use wdl_ast::Node;
 use wdl_ast::Severity;
 use wdl_doc::document_workspace;
-use wdl_engine::run;
-use wdl_engine::validate_inputs;
 use wdl_format::Formatter;
 use wdl_format::element::node::AstNodeFormatExt as _;
 use wdl_lint::rules::ShellCheckRule;
